bench/benchprof: add benchprof package
The benchprof package has been created to provide utilities for creating
CPU and memory profiles in benchmarks.

Release note: None
