+++ pkg/bench/benchprof/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "benchprof",
+    srcs = ["benchprof.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/bench/benchprof",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/testutils/sniffarg",
+        "@com_github_google_pprof//profile",
+    ],
+)
